SPI

Documentation for SPI.

SPI.adjustedrandindex

SPI.calc_spi_mtx

SPI.calc_spi_trace

SPI.calc_spi_tree

SPI.empiricalMI

SPI.getintervals

SPI.minspaceneeded

SPI.nwstr

SPI.pairwise

SPI.projectinLSV

SPI.projectinRSV

SPI.projectout

SPI.readphylip

SPI.scaledcumsum

SPI.spimtx_spaceneeded

SPI.spiresult

SPI.vmeasure_homogeneity_completeness

SPI.zscore

SPI.adjustedrandindex — Method

adjustedrandindex(a::AbstractVector{<:Number}, b::AbstractVector{<:Number}; nbins=50)

Args:

a, vector of numbers

b, vector of numbers

nbins, for continuous approximates discrete, for discrete choose nbins>maxnumberof_classes

SPI.calc_spi_mtx — Method

calc_spi_mtx(A::AbstractMatrix; [Nsmps=size(A,1), Nfeats=size(A,2), alpha=1.5, q=.75])
calc_spi_mtx(usv::SVD; [Nsmps=size(A,1), Nfeats=size(A,2), alpha=1.5, q=.75])
calc_spi_mtx(usv::SVD[, SPI.LSVs(); Nsmps=size(A,1), Nfeats=size(A,2), alpha=1.5, q=.75])
calc_spi_mtx(usv::SVD[, SPI.RSVs(); Nsmps=size(A,1), Nfeats=size(A,2), alpha=1.5, q=.75])

computes the cumulative spectral residual distance for spectral phylogenetic inference

(∑_{p ∈ P} ||UₚΣₚ||₂)²

where $P$ are the spectral partitions found with getintervals.

Args:

A,usv = AbstractMatrix or SVD factorization (AbstractMatrix is just passed to svd() before calculation)

SPI.Left() computes SPI matrix for LSVs; SPI.Right() computes SPI matrix for RSVs

alpha, q are passed to getintervals() see its documentation

Returns:

distance matrix

SPI.calc_spi_trace — Method

calc_spi_trace(usv::SVD; alpha=1.5, q=.75)
calc_spi_trace(usv::SVD[, taxaidxs]; alpha=1.5, q=.75)

calculates spectral residual within each partition of spectrum and each pair of taxa

if taxaidxs are provided the U matrix is subset and/or reordered based on those indices.

returns matrix where rows are spectral partitions and columns are taxa:taxa pairs ordered as the upper triangle in rowwise order.

SPI.calc_spi_tree — Method

calc_spi_tree(A[, ids; labelinternalnodes=true])

helper function that immediately returns the newick tree string inferred by SPI

SPI.empiricalMI — Method

empiricalMI(a::AbstractVector{<:Float}, b::AbstractVector{<:Float}[, nbins=50, normalize=false])

computes empirical MI from identity of $H(a) + H(b) - H(a,b)$. where $H := -sum(p(x)*log(p(x))) + log(Δ)$ the $+ log(Δ)$ corresponds to the log binwidth and unbiases the entropy estimate from binwidth choice. estimates are roughly stable from $32$ ($32^2 ≈ 1000$ total bins) to size of sample. going from a small undersestimate to a small overestimate across that range. We recommend choosing the sqrt(mean(1000, samplesize)) for nbins argument, or taking a few estimates across that range and averaging.

Args:

a, vecter of length N

b, AbstractVector of length N

nbins, number of bins per side, use 1000 < nbins^2 < length(a) for best results

normalize, bool, whether to normalize with mi / mean(ha, hb)

Returns:

MI

SPI.getintervals — Method

getintervals(S::AbstractVector{<:Number}; alpha=1.5, q=.75)

finds spectral partitions. Computes log difference between each subsequent singular value and by default selects the differences that are larger than 1.5 * Q3(differences)

i.e. finds breaks in the spectrum that explain smaller scales of variance

Args:

S = singular values of a SVD factorization

alpha = scalar multiple of q

q = which quantile of log differences to use; by default Q3

Returns:

AbstractVector{UnitRange} indices into S corresponding to the spectral partitions

SPI.minspaceneeded — Method

minspaceneeded(n, p; bits=64) = Base.format_bytes(binomial(n,2) * p * bits)

how much memory is needed to store spectral residual trace

SPI.nwstr — Method

nwstr(hc::Hclust[, tiplabels::AbstractVector[<:String]])

convert Hclust to newick tree string Args:

hc, Hclust object from Clustering package

tiplabels, AbstractVector{<:String} names in same order as distance matrix

SPI.pairwise — Method

pairwise(func::Function, m::M) where M<:AbstractMatrix

returns upper offdiagonals of res[k] = func(i, j) where (k, (i,j)) are calculated from enumerate(combinations(1:size(m,2), 2))

SPI.projectinLSV — Method

projectinLSV(data::AbstractArray{T}, usv::SVD{T}, [window])

returns estimated left singular vectors (aka: LSV or Û) for new data based on already calculated SVD factorization

SPI.projectinRSV — Method

projectinRSV(data::AbstractArray{T}, usv::SVD{T}, [window])

returns estimated transposed right singular vectors (RSV or V̂ᵗ) for new data based on already calculated SVD factorization

SPI.projectout — Method

projectout(usv::SVD, [window])

recreates original matrix i.e. calculates $UΣV'$ or if window is included creates a spectrally filtered version of the original matrix off of the provided components in window.

SPI.readphylip — Method

readphylip(fn::String)

Read phylip alignment file, return dataframe of IDs and Sequences

SPI.scaledcumsum — Method

scaledcumsum(c; dims=1)

cumsum divided by maximum cumulative value

SPI.spimtx_spaceneeded — Method

spimtx_spaceneeded(n, p; bits=64) = Base.format_bytes(binomial(n,2) * p * bits)

how much memory is needed to store SPI distance matrix

SPI.spiresult — Method

spiresult(A::AbstractMatrix{<:Number})

convenience function for optaining SVD, SPImtx, and SPItree

SPI.vmeasure_homogeneity_completeness — Method

vmeasure_homogeneity_completeness(labels_true, labels_pred; β=1.)

calculates and returns v-measure, homogeneity, completeness; similar to f-score, precision, and recall respectively

Args:

β, weighting term for v-measure, if β is greater than 1 completeness

is weighted more strongly in the calculation, if β is less than 1, homogeneity is weighted more strongly

Citation:

A. Rosenberg, J. Hirschberg, in Proceedings of the 2007 Joint Conference

on Empirical Methods in Natural Language Processing and Computational Natural Language Learning (EMNLP-CoNLL) (Association for Computational Linguistics, Prague, Czech Republic, 2007; https://aclanthology.org/D07-1043), pp. 410–420.

SPI.zscore — Method

zscore by columns by default. set dims=2 for rows